feat(core): normalize subtemplate resolution and uikit3->uikit folders
- PluginsubtemplateField: drop empty JDEFAULT option so a real folder is selected
- PluginLayoutTrait: resolve subtemplate via plugin param -> global com_j2commerce
  config -> bootstrap5/first-folder fallback, with existence guard
- Rename component frontend subtemplate folders + checkout CSS uikit3 -> uikit
- Flip framework token emitters (component tmpl + mod_j2commerce_cart) to uikit;
  consumers (attributes/coupon/voucher/missing_template) accept both for back-compat
- Add 6.3.7 update SQL to migrate stored uikit3 -> uikit
