+++ components/com_j2commerce/tmpl/checkout/uikit/default_register.php
