+++ components/com_j2commerce/tmpl/confirmation/uikit/default_noorder.php
