+++ modules/mod_j2commerce_cart/tmpl/detailcartonhover_uikit.php
@@ -166,7 +166,7 @@ $panelId = 'j2commerce-cart-detail-' . $moduleId;
                                         'item'       => $item,
                                         'context'    => 'cart_module',
                                         'variant'    => 'compact',
-                                        'framework'  => 'uikit3',
+                                        'framework'  => 'uikit',
                                     ], JPATH_ROOT . '/components/com_j2commerce/layouts'); ?>
                                 </div>
                             <?php endif; ?>
